+++ lib/cli/index.js
@@ -82,7 +82,10 @@ import {
 import { mergeHostInventoryBoms } from "../helpers/hostTopology.js";
 import { thoughtLog } from "../helpers/logger.js";
 import { enrichComponentWithMcpMetadata } from "../helpers/mcp.js";
-import { createNpmWorkspacePurl } from "../helpers/npmutils.js";
+import {
+  createNpmWorkspacePurl,
+  isPkgTypeOnlyImport,
+} from "../helpers/npmutils.js";
 import { isPyLockFile } from "../helpers/pylockutils.js";
 import {
   buildDependencyTrackBomPayload,
@@ -1087,47 +1090,6 @@ function hasNpmBinCommandEvidence(allImports, pkg) {
     }
   }
   return false;
-}
-
-/**
- * Helper to check if a package is imported only for TypeScript types.
- */
-function isPkgTypeOnlyImport(allImports, group, name) {
-  if (!allImports) {
-    return false;
-  }
-  const pkgNames = [];
-  if (group) {
-    const cleanGroup = group.startsWith("@") ? group : `@${group}`;
-    pkgNames.push(`${cleanGroup}/${name}`);
-    pkgNames.push(`${group}/${name}`);
-  } else {
-    pkgNames.push(name);
-  }
-
-  let hasImports = false;
-  for (const importName of Object.keys(allImports)) {
-    const isMatch = pkgNames.some(
-      (pkgName) =>
-        importName === pkgName || importName.startsWith(`${pkgName}/`),
-    );
-    if (isMatch) {
-      const occurrences = allImports[importName];
-      if (occurrences) {
-        const items =
-          occurrences instanceof Set ? Array.from(occurrences) : occurrences;
-        if (Array.isArray(items) && items.length > 0) {
-          hasImports = true;
-          if (items.some((occ) => !occ.isTypeOnly)) {
-            return false;
-          }
-        } else if (occurrences === true) {
-          return false;
-        }
-      }
-    }
-  }
-  return hasImports;
 }
 
 /**
