+++ lib/helpers/npmutils.js
@@ -249,3 +249,44 @@ export function hydrateNpmNodePackage(node, options = {}) {
     return { nodePackage, diskPkg: undefined, packageJsonPath };
   }
 }
+
+/**
+ * Helper to check if a package is imported only for TypeScript types.
+ */
+export function isPkgTypeOnlyImport(allImports, group, name) {
+  if (!allImports) {
+    return false;
+  }
+  const pkgNames = [];
+  if (group) {
+    const cleanGroup = group.startsWith("@") ? group : `@${group}`;
+    pkgNames.push(`${cleanGroup}/${name}`);
+    pkgNames.push(`${group}/${name}`);
+  } else {
+    pkgNames.push(name);
+  }
+
+  let hasImports = false;
+  for (const importName of Object.keys(allImports)) {
+    const isMatch = pkgNames.some(
+      (pkgName) =>
+        importName === pkgName || importName.startsWith(`${pkgName}/`),
+    );
+    if (isMatch) {
+      const occurrences = allImports[importName];
+      if (occurrences) {
+        const items =
+          occurrences instanceof Set ? Array.from(occurrences) : occurrences;
+        if (Array.isArray(items) && items.length > 0) {
+          hasImports = true;
+          if (items.some((occ) => !occ.isTypeOnly)) {
+            return false;
+          }
+        } else if (occurrences === true) {
+          return false;
+        }
+      }
+    }
+  }
+  return hasImports;
+}
